renamed lambda to _lambda (reserved word)
