+++ aws_operators/__init__.py
@@ -1,6 +1,6 @@
 from airflow.plugins_manager import AirflowPlugin
 from aws_operators.operators.redshift_operators import ExecuteRedshiftQueryOperator, ExecuteCopyToRedshiftOperator
-from aws_operators.operators.lambda_operators import ExecuteLambdaOperator
+from aws_operators.operators._lambda_operators import ExecuteLambdaOperator
 
 
 class AWSOperatorsPlugin(AirflowPlugin):
